+++ public/manifest.json
@@ -42,14 +42,14 @@
             },
             "description": "Toggle floating iframe"
         },
-        "toggle-modal": {
+        "toggle_modal": {
             "suggested_key": {
                 "default": "Alt+O",
                 "mac": "Alt+O"
             },
             "description": "모달 열기/닫기"
         },
-        "toggle-all-features": {
+        "toggle_all_features": {
             "suggested_key": {
                 "default": "Alt+A",
                 "mac": "Alt+A"
